fix: update implementaion of getReactFiberFromNode to be compatible with React 17
