+++ packages/fluentui/react-northstar/test/specs/utils/getReactFiberFromNode-test.tsx
@@ -1,0 +1,29 @@
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
+
+import { getReactFiberFromNode } from 'src/utils/getReactFiberFromNode';
+
+describe('getReactFiberFromNode', () => {
+  it('returns a Fiber for DOM node', () => {
+    const rootEl = document.createElement('div');
+    document.body.appendChild(rootEl);
+
+    ReactDOM.render(<div id="foo" />, rootEl);
+
+    const renderedEl = document.querySelector('#foo');
+    const fiber = getReactFiberFromNode(renderedEl);
+
+    expect(renderedEl).toBeInTheDocument();
+
+    expect(fiber).toHaveProperty('tag', 5);
+    expect(fiber).toHaveProperty('type', 'div');
+    expect(fiber).toHaveProperty('pendingProps', expect.objectContaining({ id: 'foo' }));
+
+    ReactDOM.unmountComponentAtNode(rootEl);
+    document.body.removeChild(rootEl);
+  });
+
+  it('returns null if nothing was passed', () => {
+    expect(getReactFiberFromNode(undefined)).toBe(null);
+  });
+});
